Ticket #5690:
    http://rt.dcache.org/Ticket/Display.html?id=5690

describes how dcap library has code of the form:

    sprintf(s, "%s ...", s, ...);

Ostensibly, this code will append some text to the end of s; however, the code 
is broken.  The C language specification says that the, for s*printf family of 
functions, the output buffer must not overlap with the inputs.  If they do the 
result is undefined.

For many platforms, the above construct would work as expected: the string s 
would be appended.  However, when gcc is invoked with -Wp,-D_FORTIFY_SOURCE=2 
flag and -O2 then the above construct produces unexpected results.  These 
options for gcc are those used when creating official Fedora packages.

This patch removes the reliance on this non-portable behaviour.  It introduces 
two new functions: dc_safe_strncat and dc_snaprintf.

dc_safe_strncat is similar to strncat except that the size parameter refers to 
the size of the destination string's storage.  A canonical usage is:

    char dst [100];

    dc_safe_strncat( dst, sizeof(dst), src);

This will append dst with src without overflowing dst.  dst is expected to be 
'\0'-terminated on entering the function and will be '\0'-terminated when the 
function returns.

dc_snaprintf() is a variation on snprintf().  The parameters have the same 
meaning as snprintf() but the string resulting from the formatting is appended 
on the end of the first argument string rather than replacing any existing 
content.  Thus a call like:

  char dst [100];
  strcpy( dst, "Example:");
  dc_snaprintf( dst, sizeof(dst), " %d", 10);
  printf( "%s\n", dst);

will produce output:

  Example: 10

In addition to removing the dependency on sprintf for appending a string, the 
code also removes the risk of overflowing the string buffer when building the 
dcap command.

The patch also removes a bug in dcap_poll.c where a poll event tagged 
POLLRDBAND would be processed twice.

Patch: http://rb.dcache.org/r/2027/
Acked-by: Thomas
Acked-by: Tigran
Target: trunk