Xrootd-mover: Disconnect idle clients with open files after timeout
This patch introduces an IdleStateHandler into the Netty pipeline of the
xrootd pool request handler. This IdleStateHandler will generate idle
events if neither reads nor writes have happend for the configured client
idle time period.

If such an event is received by the pool request handler, it checks if the
client has opened any files. If it hasn't, the connection is torn down.

One additional thread for all pool connections will be needed with this
patch, since the HashedWheelTimer that is keeping track of the
connections' idle state will require a thread.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Patch: http://rb.dcache.org/r/2310