pool: Improve concurrency of CacheRepositoryV5
Most of CacheRepositoryV5 is synchronized. To improve concurrency this
patch replaces a couple of hash maps with concurrent hash maps and
removes synchronization on a few methods.

Among these is getEntry() which as part of the constructor of
CacheEntryImpl extracts StorageInfo from a meta data record; when used
with the BerkeleyDB backend, this operation causes a read from the DB.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2271/