srm with terracotta: make AtomicCounter _stateChangeCounter transient
With current code in 1.9.10rc the terracotta managed srm fails with exception:
com.tc.exception.TCNonPortableObjectError:
*******************************************************************************
Attempt to share an instance of a non-portable class referenced by a portable class. This
unshareable class has not been included for sharing in the configuration.

For more information on this issue, please visit our Troubleshooting Guide at:
http://www.terracotta.org/kit/reflector?kitID=default&pageID=npoe

Referring class        : org.dcache.srm.request.PutRequest
Referring field        : org.dcache.srm.request.ContainerRequest._stateChangeCounter
Non-portable field name: org.dcache.srm.request.ContainerRequest._stateChangeCounter
Thread                 : init
JVM ID                 : VM(115)
Non-included class     : org.dcache.commons.util.AtomicCounter
...

Exception is much larger, but it pretty much means that we try to save in Terracotta managed memory a field of the type that is not registered with Terracotta.  This new type is org.dcache.commons.util.AtomicCounter, which was added by Gerd to implement the delay of switch from synchronous to asynch ronousexecution mode for srm requests. The field of this type, org.dcache.srm.request.ContainerRequest._stateChangeCounter, does not need to be saved in the shared memory as it is only used during the original submission, so instead of registring it with terracotta, we mark it "transient".