pool: Refactor MetaDataRecord cache logic
CacheRepositoryV5 loads all MetaDataRecords into a cache during
initialization.

This patch refactors the logic such that the cache is implemented by the
MetaDataCache decorator. MetaDataCache supports concurrent read of
entries from the decorated MetaDataStore. This will serve as the
foundation of future patch that will allow reads of entries even before
the repository is fully initialized.

The change adds several new failure modes to Repository operations.
Additional exceptions are added to the throws declaration of Repository
methods, and the rest of the pool code is updated to handle those
exceptions. Many of these failures are unlikely to happen with the
current cache implementation, as entries will stay in memory once read.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2291/