PinManager: Use toplink.application-location property
Define the location of the toplink DDL files by using the
toplink.application-location property. Move the property definition to the
batch file so it can be based on the respective path property from
paths.property. Use toplink.create-ddl-jdbc-file-name only for the name of
the file.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Target: trunk
Patch: http://rb.dcache.org/r/2365