srm: add options for storing srm request in db only once upon completion and not using database for request storage at all The new options are srmJdbcEnabled (default true) and srmJdbcSaveCompletedRequestsOnly (default false) this is accomplished by the usage of the two new implementations of the JobStorage interface and corresponding JobStorageFactories. If the srmJdbcEnabled is set to false then srm will use an instance of the NOOP Job Storage. If the srmJdbcEnabled and the srmJdbcSaveCompletedRequestsOnly is set to true, then the FinalStateOnlyJobStorageDecorator will be used to decorate current Jdbc Job Storages. FinalStateOnlyJobStorageDecorator only propagates the save call to the underlying storage iff the job is in the final (Failed, Done or Canceled) state.
Motivation for this patch:
I initially thought of storig the requests in database for 
1. Monitoring purposes
2. Sharing state between multiple instances of srm
3. Ability to restore state after restart

It now appears that sharing state can be done more efficiently using other technologies (i.e. Terracotta, JMS, etc), and clients usually do not survive the restart of the server, so we only really need database for monitoring purposes and sites that do not rely on SRMWatch, do not really need requests in database at all.
This patch will allow the sites using srm posibilities to store each request in db only once. This will make it impossible to see the requests in progress though SRMWatch, but will reduce the load on postgres significantly, and others who want even more lightwaitgh srm can just turn using db off completely. SRM will still rely on postgress for generation of the request ids and for storage or the request credentials, but this is much smaller load on the database.