pool: Enable pool readonly while loading entries
This patch enables the pool in read-only mode before the complete
repository is loaded from disk. The change required the Repository
interface to be changed to dinstinguish between the initial listing of
the data directory and the loading of the entries.

This patch changes CacheRepositoryV5 such that it allows to call
openEntry, getEntry, getState and setSticky while the repository is
still loading.

The patch changes PoolV4 such that the repository is loading in a 
background thread. This allows severals pools in the same domain to
be loaded concurrently.

Unit tests have been updated to use the new interface.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2377/