dcap : memory issues Protection against message truncation
Charles Waldman was getting messages of character lengths in the region of 3220. Along with messages from dcache saying missing parameters.

http://rt.dcache.org/Ticket/Display.html?id=5774

I looked at the previous patch, http://rb.dcache.org/r/2027 which addressed the issue of over running the buffer, but did not solve the message truncation issue. This patch solves both the issues solved in http://rb.dcache.org/r/2027

Like the previous patch I based this one on the error handling from the man page from snprintf, I also solve the issue that the C language specification says that the, for s*printf family of functions, the output buffer must not overlap with the inputs.  If they do the result is undefined.

Signed-off by: Paul and Tigran
http://rb.dcache.org/r/2356/