PoolManager: Implement rebalance command
This patch adds a first attempt at a simple rebalance command. It allows
one-time rebalancing of a pool group. The command is implemented on top
of the migration module. The pool manager part is stateless; it simply
generates a set of migration commands which are submitted to the pools
involved.

To avoid that more than one rebalancing job is running on a pool, the
patch makes it possible to add a custom ID to a migration job. Before
submitting a new job, the PoolManager cancels any previous rebalancing
job on the pool.

The current implementation is pretty basic; it only implements balancing 
of pool groups, doesn't have any query functions, and only implements 
two balancing metrics. Those are things we can add over time.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2395/