cells: Change error handling for String commands
If an ac_ method throws an exception, cells by default converted the
exception to a String and returns that String to the requestor. This
works if the requestor is a human being, but for other cells this
behaviour forces them to parse the reply string to detect errors.

This patch changes the default such that cells propagates the exception.
The ssh shell converts the exception to a string, so for a human being
it should not make much of a difference.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Patch: http://rb.dcache.org/r/2392/