srm: Move file locality logic to PoolManager
This is a follow up to review request 2421.

To avoid the cost of querying the pool for information the PoolManager
already has, this patch moves the file locality logic to PoolManager.
This also allows us to use additional information about the PSU in
future patches. For now the logic is more or less the same as the one
introduced by patch 2421.

There is one crucial difference in the logic between 2421 and this
patch: 2421 only used the pool mode for detecting disabled pools.
However a pool can also be disabled in the pool selection unit. The
present patch will take that into account. I.e. 'psu set pool pool_0
disable' will cause files on pool_0 to become UNAVAILABLE.

Target: trunk
Target: 1.9.10
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/2447/