dcap: fix compilation on Solaris
There were two problems when trying to compile dcap for Solaris that a previous 
patch introduced.

The first problem is trivial: the man-page for the stat(2) command states that 
sys/types.h, sys/stat.h and unistd.h must be included.  However, it doesn't state in 
which header file struct stat is defined.

The dcap.h file includes function prototypes using struct stat.  It includes 
sys/types.h, but not sys/stat.h.  On GNU systems, including sys/types.h is 
sufficient for a definition of struct stat; however, on Solaris machines sys/stat.h 
must be included for the definition of struct stat.

The fix is to include sys/stat.h in dcap.h

The second problem is slightly more complex: on Solaris, the stat(2), lstat(2) and 
fstat(2) methods are defined, in the header sys/stat_impl.h, as static local 
functions.  This means that each compiled object file will include its own 
implementation of the stat, lstat and fstat functions.  These implementations are 
simply a wrapper around "x"-versions of the functions (_xstat, _lxstat, _fxstat 
respectively).  The purpose is to encode a hard-coded version number as an argument 
to these methods as a backwards compatibility measure.

However, this means that, in dcap_preload.c, there are two definitions for each of 
the three stat-functions (stat, lstat, fstat).  The first set of definitions come 
from sys/stat_impl.h and the second set comes from dcap_preload.c itself.  The 
proposed solution is to move the definition of stat, lstat and fstat into the 
corresponding #else clause.  This allows the "built-in" wrapper in Solaris to 
operate, mapping (for example) stat --> _xstat.  The _xstat implementation is 
overloaded by dcap_preload.c, which will handle the request using dcap's 
implementation of stat.

Patch: http://rb.dcache.org/r/2449/
Target: trunk
Acked-by: Tigran