dcap: fix Solaris building (second part)
The file debug_level.h has a dual purpose: it contains the definitions of the 
various debug levels and it also is the entry-point for the HAVE_ macros that the 
autoconf system builds.

In future, these two purposes need to be separated; however, for now, we fix the 
compilation problem by accepting the dual-role of this file and including it when we 
need to know the value of the HAVE_ macros.

Patch: http://rb.dcache.org/r/2451
Acked-by: Tigran
Target: trunk