dcap: fix bug in dc_fgets
The function dc_fgets is the dcap wrapper of the POSIX fgets.  The implementation 
contains a bug: if the size argument is zero then the return value depends on an 
uninitialised variable: n.  If the uninitialised value of n is less than zero then 
the function returns the wrong value.

The patch fixes this by initialising n to zero.

Patch: http://rb.dcache.org/r/2443/
Target: trunk
Acked-by: Owen