dcap: fix dc_writev bug
The dc_writev function wraps the POSIX writev vector write (a.k.a "gather write") 
function.  It loops over the vector and builds a complete buffer that is sent to the 
mover.

The loop depends on the variable iobuf_pos, which is incremented when processing the 
elements of the vector; however, iobuf_pos is not initialised to zero so may take an 
arbitrary value.

This patch fixes this by initialising iobuf_pos to zero.

Target: trunk
Patch: http://rb.dcache.org/r/2444/
Acked-by: Owen