srm: Fix ls compatibility with old dcache srm clients
Old dCache SRM clients (and the one we have for download on our home
page is old) have a bug in the ls code: To disinguish between
synchronous and asynchronous ls responses, they don't look at the return
status but at the request token: if one is included the client assumes
we are looking at an asynchronous reply.

Our current trunk code always includes a request token in the reply, no
matter whether it is synchronous or asynchronous. That's nice for
debugging and tracability, however it breaks our own client. The client
bug has been fixed on trunk, however all deployed copies of the client
still have the bug.

Therefore this patch changes the server to only include a request token
in asynchronous ls replies.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2463/