scripts: fix chimera-cli.sh
The chimera-cli.sh command currently doesn't work in trunk (and 1.9.10).  The 
problem is that it tries to run the support library (loadConfig.sh) rather 
than "sourcing" it.

This patch addresses the issue in RT ticket: 5878

    http://rt.dcache.org/Ticket/Display.html?id=5857

This patch fixes the problem by sourcing the support library rather than 
attempting to run it.

Patch: http://rb.dcache.org/r/2461/
Acked-by: Dmitry
Target: trunk
Request: 1.9.10