dcap: fix (void *) pointer arithmetic in dcap_read
The buffer argument to dcap_real_read function is defined to be (void *), i.e., a 
pointer to some absolute location in memory.

However, the dcap_real_read implementation treats the buffer pointer as if it 
pointed to an array of bytes, which is natural.  However, in doing this, the code 
increments the pointer as data is read.  This doesn't make sense for a (void *) 
pointer, it only makes sense for a pointer to an array of some type with 
well-defined length.

This patch fixes this by accepting a (void *) pointer as an argument but 
assigning the value to a (char *) variable.  This makes explicit the semantics 
currently only tacit in the code.

Patch: http://rb.dcache.org/r/2460/
Target: trunk
Acked-by: Gerd