dcap: fix possible use of uninitialised variable
The copyfile method in dccp is responsible for copying the data from source to 
destination.  This function allocates a buffer to hold an amount of data.  If 
this allocation fails then the file size is not set and the corresponding print 
statements use an uninitialised value for the transferred size.

This patch fixes this by initialising size to zero.

Target: trunk
Patch: http://rb.dcache.org/r/2457/
Acked-by: Gerd