dcap: fix type casting of type-punning
In dcap_error.c, the support for thread-local variable, t_getspecific macro, 
which is a simple wrapper around pthread_getspecific(), is used in dcap_error to 
obtain a ptr to the thread-local errno.  The dcap_error casts this pointer to a 
(void *), rather than (int *).

Assigning to a type-punning, like this, is not allowed by C89/C90 standard but 
historically, compilers turn a blind eye to them.

It's legal to assign a (void *) point to any other kind of pointer without 
casting.  The pthread_getspecific() function returns a (void *) pointer, so the 
explicit casts in dcap_error are actually unnecessary.

The proposed patch fixes the "dereferencing type-punned pointer will break 
strict-aliasing rules" warnings.

Patch: http://rb.dcache.org/r/2452/
Target: trunk
Acked-by: Gerd