cells: Fix help message regression
It used to be the case that typing an invalid or incomplete command
would output a help text together with the error message. This
functionality was lost when we changed cells to return an exception in
case of errors rather than a String.

This patch restores the lost functionality.

Target: trunk
Target: 1.9.10
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2469/