dcap: add -H option to dccp that provides feedback during transfer
Whilst in Karlsruhe for GridKa school, I was talking to someone about their experience of dCache 
being very slow when copying files into and from dCache via dccp.  The user was unable to answer 
simple questions that might indicate what is the problem (where you waiting for a mover? was the 
checksum taking a long time? was the transfer slow on-average, or did it "stutter"?).  I released 
that this information is currently difficult (or impossible) to obtain with dccp.  Moreover, without 
any feedback during a transfer, one must have blind faith that the transfer is proceeding.

This patch addresses this issue.  It provides some feedback to the user during a transfer.  The 
feedback uses the (old) method of writing output without the line-feed character ('\n') but using the 
carriage-return character ('\r') instead.  Repeated use of such output gives the user the impression 
of a single line of text is being updated with new information.  Some care is needed to append spaces 
to the end of the new version of the line if it has shrunk.  This is to hide character(s) from the 
previous line.

The transfer feedback line contains a progress bar followed by zero or more items.  The items are: 
percentage of file transferred, number of bytes transferred and the average transfer rate.  The code 
determines the width (number of columns) of the terminal dynamically and scales the output 
accordingly.  If the terminal has sufficient width then all items are displayed and the bar is scaled 
to take (almost) all of the line's remaining capacity.  If the width is too small to allow this then 
items are progressively switched off to provide additional space, allowing the progress bar to grow 
in size.  This continues until a minimum width of the progress bar is achieved or all items are 
switched off (whichever happens first).

The progress bar is a sequence of characters the first being the open square-bracket and the last 
being the close square-bracket.  Those in between are either a dash ('-') or a hash ('#') and form 
the part that indicates the transfer's progress.

If the size of the file being transferred is known in advance, the progress bar is a "percentage 
progress bar".  This shows information as a sequence of zero or more hash symbols followed by zero or 
more dashes.  The total number of symbols is fixed, for given a progress bar size. The number of 
hash- and dash- symbols is chosen to be proportional to the amount of data written and still to write 
respectively.  The result is an ASCII-art version of the familiar GUI progress-bar.  It is also how 
the rpm command shows progress.

The following is an example line when the file size is known and the terminal is sufficiently wide to 
allow all items:

    [#######---------------------------------------------------------------] 11% 45.0 MiB (4.09 MiB/s)

If the size of the transfer isn't known in advance (e.g., data is taken from stdin) then it is 
impossible to show a "percentage activity bar".  Instead, the progress bar is an "activity progress 
bar".  This indicates bytes transferred as a shorter bar (made up of some five hashes).  This shorter 
bar is initially left-most within the progress bar and "moves to the right" as data is written.  When 
the shorter bar reaches the end of the activity bar, it wraps.  (This similar to how Firefox shows 
progress under similar circumstances.)

The following is an example when the transfer size is unknown and the terminal is sufficiently wide 
to allow all items:

    [-----#####----------------------------------------------------------------] 8.00 MiB (4.00 MiB/s)

In writing this support, I adding a function for formatting file sizes using SI binary prefixes (kiB, 
MiB, etc), taking care of aesthetics.  The patch also updates the final line of output (which gives a 
summary of the transfer) to also use the new function.

The following is an example of current dccp's final output line:

    417488294 bytes in 78 seconds (5226.97 KB/sec)

This is an example of final output line after patch:

    417488294 bytes (398 MiB) in 76 seconds (5.24 MiB/s)

Current limitations and future work:
    o  no user feedback when the transfer stalls,
    o  could add an estimate of the time remaining until the transfer finishes.
 
Target: trunk
Patch: http://rb.dcache.org/r/2400/
Acked-by: Gerd
Acked-by: Owen