dcap: fix bootstrap.sh to work on Solaris
Solaris packaging provides tools like aclocal and automake as filenames with a 
version-specific suffix (e.g., aclocal-1.11, automake-1.9, etc); moreover, we cannot 
necessarily mix-and-match.  We must use tools with a consistent version; for example, 
running v1.11 of aclocal with automake.

The current implementation of bootstrap.sh simply invokes "aclocal" and "automake" 
without any version-specific suffix.  This doesn't work on Solaris.  There is 
commented-out lines in the current bootstrap.sh that include a version-specific suffix, 
suggesting this issue is known but not addressed.

This patch fixes the problem.  It looks for a set of consistently named tools with a 
version-specific prefix, in decreasing version numbers (1.11, then 1.10, then 1.9).  If 
these all fail, it tries again without any version-specific suffix.  If none of these 
succeed then the script fails with a suitable error message.

The output (on the terminal) from running bootstrap.sh should be identical to 
previously.  The behaviour on running bootstrap.sh should be identical, including that 
it terminates immediately after any failing step.

Patch: http://rb.dcache.org/r/2465/
Acked-by: Owen
Target: trunk