dcap: refactor dcap debugging slightly
dcap library provides a debugging support where different channels ("levels") of output 
may selected.

This patch adds two additional methods: dc_vdebug and dc_is_debug_level_enabled.  Both 
are exposed to other components within the dcap library but are not listed as public 
methods.

The dc_vdebug method is a var-args version of the dc_debug method.  This is similar to 
the 'v' version of the printf family of functions (printf vs vprintf, etc).

The dc_is_debug_level_enabled returns True (1) if the supplied level is enabled, False 
(0) if not.  This is intended to allow external modules to avoid work that will not be 
used in generating output.

Patch: http://rb.dcache.org/r/2510/
Target: trunk
Acked-by: Gerd