xrootd: Do not unbind server socket upon client connect
The idea of the patch in [1] was to make the xrootd mover less
resource-consumptive by terminating the accepting parent socket once the
connection with the first client was established, as it would only serve
exactly one client anyways.

However, as a side effect, this made the port on which the mover was
running available for reuse by another mover that gets started on the
same pool. If the same client is talking to both of the movers (for
instance when opening a TChain in ROOT), it will assume that both open
requests use the same TCP connection. It will send its open request to
the mover with which it first established the TCP connection (the first
mover), which will refuse to accept it as the checksum of the second
open request does not match the checksum it received from the door
(which is the checksum of the first open request).

To avoid that erroneous behavior, this patch reverts the patch in [1].
All supported releases from 1.9.5 to 1.9.9 are affected.

[1] http://rb.dcache.org/r/2150/
Patch: http://rb.dcache.org/r/2526/
Signed-off-by: Thomas Zangerl