dcap: update test program
The program dcap_test provides a basis for testing the dcap library.  It uses a 
specially formatted file, creating it if necessary, where the file contains offset 
information stored as 4-byte integer values.  This allows the test to verify the 
contents of the data when reading.

The patch alters the format of the data in two ways:

    1) the byte-order of the integers is fixed to be
       network-byte-order.  Previously it was host byte-order,
       which lead to ambiguity that had to be worked around.

    2) the offset had a somewhat counter-intuitive meaning:
       the offset recorded in the file is the offset of the
       following byte.

After the patch, the integer value of a 4-byte sequence is the offset of the first 
of the four bytes.  The first 32 bytes of the file are now:

    00 00 00 00 00 00 00 04   00 00 00 08 00 00 00 0c
    00 00 00 10 00 00 00 14   00 00 00 18 00 00 00 1c

The dcap_test program tests the ability to read data by repeatedly reads a (fixed) 
block of data and skipping forward a small, random amount until the file is 
exhausted.  The data read is verified by checking it against the expected values.

This patch enhances the tests.  It decouples the access patterns from the mechanism 
for running the pattern.  In addition to the existing access pattern (now called 
"walking forward"), the patch introduces additional access-patterns: "walking 
forward with random initial offset", "walking backward", "random 4-byte reads" and 
"random block reads".

The patch includes one "runner" that implements a round-robin selection of the 
selected access patterns.  Any number of access patterns may be given.  This allows 
multiple patterns to be stacked; for example, "walking forward" + "walking backward" 
gives a cross-like access pattern.

The access-patterns are written to externalise their state, so multiple instances of 
the same access-pattern may be used by a runner.  Combining two "walking forward 
with random initial offset" simulates the ATLAS ROOT access pattern.

The round-robin runner has a selectable terminate conditions.  There are three 
possible termination conditions: after the first access-pattern completes, after the 
last access-pattern completes, or after a fixed number of iterations.  If an 
access-pattern terminates before the runner terminate condition then that 
access-pattern is disabled and the runner will continue with the remaining 
access-patterns.

Possible future work is described at the beginning of the file.  These include 
adding a multithreaded runner, to test dcap when used by multiple threads 
concurrently; and adding support for testing multiple file descriptors concurrently.

Patch: http://rb.dcache.org/r/2478/
Target: trunk
Acked-by: Owen
Acked-by: Tigran