srm: remove SrmStartupPing class
The class SrmStartupPing provides a CLI that contacts the SRM service and 
issues an srmPing request.  This was needed to start the SRM as the entire 
srmDomain, which contained services like srm-space-manager, was hosted within 
Tomcat+Axis and Axis provided no standard mechanism to start a service.

With the move to Jetty, only the SRM is hosted within the Axis servlet.  The 
other components become standard services within whichever domain they should 
run and so were started up automatically.  With this, the need to "ping" the 
SRM goes away and so the SrmStartupPing class is now redundant.

This patch removes the class.

Target: trunk
Patch: http://rb.dcache.org/r/2546
Acked-by: Dmitry