srm: Refactor representation of SURLs and paths
There are currently way too many representations of SURLs and paths:

- axis URI class
- GlobusURL
- SrmUrl
- URL
- URI
- FsPath
- String
- File

and the String form covers several representations:

- Surl
- Turl
- path component of surl
- local path including the root prefix

Conversions between the representations are scattered all over the
place, with repeated checks for whether the SURLs are valid (e.g. refer
to the local site).


This patch is proposal to clean up the path representation. It is not a
complete patch; it just touches one corner to show the direction I am
proposing we move in. The patch itself arguably makes the situtation
worse by being incomplete and thus making things even more inconsistent.


The idea is simple:

- The generic SRM operates on SURLs and TURLs represented as
  java.net.URI. The use of Axis URI is necessary for compatibility with
  Axis, however the representation should as quickly as possible be
  converted to Java's standard URI class.

- The AbstractStorageElement (the backend of the SRM) operates on SURLs
  and TURLs represented as java.net.URI (just like the frontend). This
  let's the frontend part be ignorant about the mapping to local paths,
  whether a host name is local or not. It also gives more power to the
  backend by letting it operate on a larger name space of SURLs.

- diskCacheV111.srm.dcache.Storage (the SRM backend connecting to
  dCache) maps SURLs to FsPath. FsPath is a complete path in the name
  space of dCache (i.e. it includes any root prefix that may have been
  configured in the SRM door). Validity of the SURL (e.g. does it refer
  to the correct host name) is checked while converting the URI to
  FsPath.


Some complications are:

- srmCopy has a lot of logic in the SRM frontend for calling out to
  scripts to perform the copy. I don't know yet whether this is an issue
  or not.

- The srmLs response refers to the absolute path to the files,
  according to the SRM spec. The type in the protocol for the path is
  however String and not URI. That breaks the rule about always using
  URI in the frontend. I do however believe we can live with this small
  exception.


The present patch updates:

- getPutTurl, getGetTurl, getFileInfo, prepareToPut, getFileMetaData of
  AbstractStorageElement and its implementations.

- Updates handlers and request objects that use the above methods.

- Updates several request objects to use java.net.URI where possible.


One consequence of the patch is that escaped symbols (e.g. white space) 
in file names are now handled correctly. If one uses a client that can
handle this, one can upload a file that includes a space in the name:

ngcp /bin/bash srm://dcache-vm:8443//a%20d

The resulting file is

behrmann@dcache-vm:/pnfs/dcache-vm/data$ ls /pnfs/dcache-vm/data/a*
/pnfs/dcache-vm/data/a d

This doesn't work with the dCache srmclient, as the client is not able
to deal with the escaped space in the TURL. Ironically our current 
client succeeds when this patch is not applied, although with the 
wrong result:

/opt/d-cache/srm/bin/srmcp file:////bin/bash srm://dcache-vm:8443//a%20a

behrmann@dcache-vm:/pnfs/dcache-vm/data$ ls /pnfs/dcache-vm/data/a*
/pnfs/dcache-vm/data/a%20a

That's not supposed to happen. 

On the other hand, without this patch, ngcp (the ARC client) fails to
upload the file; that's because it does handle the escape sequence in 
the TURL and uses a real space. However the SRM fails to recognize it 
during srmPutDone.

If the patch gets accepted (i.e. if Dmitry is happy with the direction), 
then I will update the remaining functions in the AbstractStorageElement.


Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.org>
Patch: http://rb.dcache.org/r/2417/