srm: Refactor representation of SURLs and paths [step 4]
The previous patches changed the AbstractStorageElement interface to be
URI based. This patch updates most of the request logic in the SRM to be
URI based too and moved the conversion between Axis URI and Java URI to
the handlers.

The toUris method is repeated in several handlers. Those handlers
contain lots of repeated code already. Rather than introducing a utility
class for toUris, I suggest that at some point we refactor the handlers
to share more code in a base class.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/2450/