dcap: add initial support for local caching (dcap++)
This patch adds support for a simple LRU cache: a Local Cache Buffer (LCB).

It is a reworked version of the patch from Guenter Duckeck 
<Guenter.Duckeck@physik.uni-muenchen.de>.  When reworking it, I've tried to make 
the algorithm more readable by splitting the functionality into separate 
functions.  I've also removed code-duplication (as much as possible), those 
variables and structure members that weren't actually used, all the commented-out 
code and the (logical) dead-code.

I've tried to fix some of the issues Gerd raised in his initial review; however, 
some of these issues, whilst certainly valid, may not be worth fixing here.  
Instead, we can address them with a rewrite of dcap so it has a generic caching 
layer (as Tigran proposes).

I've copied Gerd's comments that are still outstanding into the code as comments 
prefixed with "REVISIT".

I'm aware of a number of limitations with the code as-is:
  a) the first cache block (block-number 0) must always be cached,

  b) the algorithm is O(n), where n is the number of blocks,

  c) behaviour under error conditions [either dc_real_read() fails
     returning -1 or it read fewer bytes than requested when not
     at EOF] is poor.

  d) the code disables read-ahead (RA) which, I believe isn't necessary.
     Current implementation of LCB+RA is suboptimal, but should be
     functional.

  e) if LCB+RA is not allowed, then client code can subvert this
     by explicit call to dc_setBufferSize().

Patch: http://rb.dcache.org/r/2475/
Target: trunk
Acked-by: Gerd
Acked-by: Tigran
Acked-by: Owen