build: don't build empty xrootd-tokenauthz.jar
The jar file xrootd-tokenauthz.jar is created as part of the dCache build process.  This used to 
contain dCache's support for the catalogue-based authorisation with the old xrootd door.  The new 
xrootd no longer needs this jar and now, with the removal of the underlying Java code, the jar is 
now empty.

Distributing an empty jar file doesn't make a lot of sense, so this patch adjusts the build 
process so we no longer build xrootd-tokenauthz.jar.

Target: trunk
Patch: http://rb.dcache.org/r/2553
Acked-by: Jan