ACLs: fix comparison between Boolean objects
The class AclNFSv4Matcher compares whether a Boolean variable is false by checking the value is the same 
object is Boolean.FALSE.  This is not robust as it's possible to create a new Boolean object that is 
equal to Boolean.FALSE but not the same object.

The problem is in the PDP method isAllowed with the extra "allowMask" parameter.  It appears this method 
is currently not used in dCache.

Target: trunk
Acked-by: Tanja
Patch: http://rb.dcache.org/r/2575/