login: Introduce login and group principals
This patch makes two changes to our principal classes:

- It introduces a GroupPrincipal interface shared by all group principals.

- It introduces LoginNamePrincipal, LoginUidPrincipal, and LoginGidPrincipal.

The latter three principals are similar to UserNamePrincipal,
UidPrincipal, and GidPrincipal, but with slightly different semantics:
These represented identities that have not been authenticated yet. It is
the task of the LoginStrategy to authorize use of these principals and
translate them to UserNamePrincipal, UidPrincipal and GidPrincipal.

LoginUidPrincipal and LoginGidPrincipal are not used yet, however I
execpt that the DCAP door will eventually generate these.

RELEASE NOTES: This patch requires a synchronized upgrade of all 
non-pool nodes.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2597/