srm: Fix srmLs performance regression
In 1.9.10 we changed how we compute file locality. Because of this, the
attributes required for computing file locality have changed.
Unfortunately a small bug causes us to fetch the additional attributes
even when verbose listing is not requested. This causes a performance
regression.


Target: trunk
Request: 1.9.10
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2621/