pool: make Fifo/LifoPriorityComparator serializable
as suggested by findbugs:

This class implements the Comparator interface. You should consider
whether or not it should also implement the Serializable interface. If a
comparator is used to construct an ordered collection such as a TreeMap,
then the TreeMap will be serializable only if the comparator is also
serializable. As most comparators have little or no state, making them
serializable is generally easy and good defensive programming.

Acked-by: Paul Millar
Patch: http://rb.dcache.org/r/2623/