SpaceManager: Propagate the file size to PoolManager and pool
The SpaceManager will often know the intended file size as this is part
of the srmPrepareToPut message. This information was however not
propagated to the PoolManager or the pool. The PoolManager needs to know
so that it can take the file size into account when selecting a write
pool. The pool needs to know as it can use it to allocate enough space
on the pool right away and it can use it to detect incomplete transfers.

Notice that SRM considers 0 and 1 to be special file sizes: Space
manager must not consider these as real file sizes.

RELEASE NOTES: The patch affects how we treat incomplete uploads. If a
file size is provided by the SRM client, the upload will now
automatically fail if the uploaded file's size does not match the
expected file size. Previously such uploads would succeed.

Previously submitted as: http://rb.dcache.org/r/1597/

Target: trunk
Ticket: #5075
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/2619/